apimisc_if_get_catalog_list

Retourne la liste des catalogues de l'API info fi

apimisc_if_get_catalog_list()

Valeur de retour

un data.frame du catalogue de l'API info fi

Examples

apimisc_if_get_catalog_list()
#> $total_count
#> [1] 6
#> 
#> $results
#>   visibility             dataset_id dataset_uid has_records           features
#> 1     domain sous-type-dinformation   da_muue1b        TRUE            analyze
#> 2     domain         societes-cac40   da_3il2gc        TRUE            analyze
#> 3     domain      flux-amf-new-prod   da_z9cfdi        TRUE timeserie, analyze
#> 4     domain         liste-code-isi   da_md2skn        TRUE            analyze
#> 5     domain     types-dinformation   da_goam2i        TRUE            analyze
#> 6     domain              codes-lei   da_9hhjm0        TRUE            analyze
#>   attachments alternative_exports data_visible
#> 1        NULL                NULL         TRUE
#> 2        NULL                NULL         TRUE
#> 3        NULL                NULL         TRUE
#> 4        NULL                NULL         TRUE
#> 5        NULL                NULL         TRUE
#> 6        NULL                NULL         TRUE
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     fields
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              codelov, type_d_information, sous_type_d_information, type_of_information, subtype_of_information, type_informationn, NA, TRUE, TRUE, TRUE, TRUE, TRUE, NA, NA, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, >, NA, NA, NA, NA, NA, NA, codelov, Type d'information, Sous-type d'information, Type of information, Subtype of information, type d'informationn, text, text, text, text, text, text
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      id, name, isin, rcs, cac40, name_cac40, first_letter, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, -count, -count, NA, NA, NA, NA, NA, NA, TRUE, TRUE, TRUE, ID, NAME, ISIN, RCS, CAC40, name_cac40, first_letter, int, text, text, text, text, text, text
#> 3 uin_ver_fin, uin_ver_lov, uin_idt_uin, uin_dat_amf, uin_dat_mar, filename, identificationdiffuseur_idi_cod_dif, identificationsociete_iso_nom_soc, identificationsociete_iso_cd_amf, identificationsociete_iso_url_irg, identificationsociete_iso_pay_ss, identificationsociete_iso_cd_isi, identificationsociete_iso_url_int, identificationsociete_iso_url_log, identificationsociete_iso_cd_lei, fichierdecontenu_inf_fic_nom, informationdeposee_inf_dat_emt, informationdeposee_inf_tit_inf, informationdeposee_inf_stp_pri, informationdeposee_inf_cod_dif, informationdeposee_inf_stp_inf, informationdeposee_inf_lst_dif, informationdeposee_inf_lng_inf, identificationsociete_iso_code_tkr_iso_cd_mch, identificationsociete_iso_code_tkr_iso_cd_tkr, informationdeposee_inf_upg_inf_inf_upg_rel, informationdeposee_inf_upg_inf_inf_upg_sts, url_de_recuperation, timezone_uin_dat_amf, cac40, name_cac40, type_d_information, sous_type_d_information, type_of_information, subtype_of_information, type_informationn, type_d_information_ancien, type_d_information_nouveau, type_of_information_ancien, type_of_information_nouveau, code_isin_nom_sc, UIN_VER_FIN\nN° de version de la structure de fichier du « Flux d’informations » qui correspond à la version en vigueur de l’interface, UIN_VER_LOV\nN° de version du contenu du « Référentiel LOV » chargé dans le SI du Diffuseur, UIN_IDT_UIN\nIdentifiant de l’unité d’information transmise par le Diffuseur, UIN_DAT_AMF\nDate technique correspondant à l’évènement générant l’envoi de l’unité d’information à l’AMF, UIN_DAT_MAR\nDate technique correspondant à l’évènement générant l’envoi de l’information au marché.\nHorodatage GMT.\n01/01/9999 si la diffusion au marché intervient à une date inconnue. \n01/01/8888 si l’information n’est pas diffusée vers le marché (diffusion réservée à l’AMF).\n, NA, IDI_COD_DIF\nCode LOV permettant d’identifier le Diffuseur, ISO_NOM_SOC\nNom de l’émetteur (dénomination juridique) , Code ID AMF de l’émetteur, ISO_URL_IRG\nURL de la zone du site Internet de l’émetteur dédié à l’archivage de l’information réglementée, ISO_PAY_SS\nPays du siège social de l’émetteur, ISO_CD_ISI\nCode ISIN de la ligne de cotation principale de l’émetteur ou à défaut d’une ligne secondaire , ISO_URL_INT\nURL du site Internet de l’émetteur, ISO_URL_LOG\nURL vers le logo de l’émetteur, NA, INF_FIC_NOM\nNom du fichier de contenu attaché (syntaxe : ‘FC’ + « IDI_COD_DIF » + « UIN_IDT_UIN » + ‘.xxx’), INF_DAT_EMT\nDate technique de transmission de l’unité d'information au Diffuseur par l’émetteur, INF_TIT_INF\nIntitulé de l’information déposée par l’émetteur, INF_STP_PRI\nCode LOV du type d’information déposée par l’émetteur, considéré comme principal.\n\nUne des valeurs renseignées dans le champ INF_STP_INF est désignée en tant que type d’information principal. C’est le type qui porte le sens de l’information déposée. \nSi une seule valeur est renseignée dans INF_STP_INF, alors INF_STP_PRI  = INF_STP_INF. \n, Code précisant la règle de diffusion de l'unité d'information au marché\net à J'AMF (dépend du type d'information principal\nINF _STP _PR!. cf. LOV STYP du « Référentiel LOV »), INF_STP_INF\nCodes LOV du type d’information déposée par l’émetteur.\n\nChamp multivalué.\nCodification administrée par l’AMF, mise à disposition dans le Référentiel LOV.\nLa valeur INF_STP_PRI doit y être incluse.\n, INF_LST_DIF\nListe de diffusion qui précise les destinataires de l’information (dont l’AMF) et en particulier les médias situés en France ou dans les pays de l’Union européenne qui reprennent l'information diffusée en version brute et intégrale. , INF_LNG_INF\nLangue de l’information déposée par l’émetteur, Code pays du marché relatif au code Ticker transmis dans le champ ISO_CD_TKR, Code Ticker de la ligne de cotation principale de l’émetteur ou à défaut d’une ligne secondaire., Nom du fichier d’une information antérieure liée (syntaxe : ‘MD’ + « IDI_COD_DIF » + « UIN_IDT_UIN » + ’.xml’)\n\nChamp lié au INF_UPG_STS.\nRenseigné obligatoirement si INF_UPG_STS <> ‘NEW’ (cf. LOV STTS du « Référentiel LOV ») sinon vide.\n\n\n, INF_UPG_STS\nCode LOV du statut précisant si l’information déposée correspond à un additif ou rectificatif d’une information antérieure , NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, TRUE, NA, NA, TRUE, NA, NA, TRUE, NA, NA, NA, NA, NA, TRUE, NA, NA, NA, NA, TRUE, TRUE, NA, NA, NA, TRUE, NA, NA, TRUE, TRUE, TRUE, TRUE, NA, NA, TRUE, TRUE, TRUE, TRUE, TRUE, NA, NA, NA, NA, TRUE, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, hour, hour, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, hour, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, -alphanum, -count, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, -count, NA, NA, NA, NA, NA, NA, NA, NA, NA, -count, -count, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, NA, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ,, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >, NA, NA, NA, NA, NA, Version flux information, Version LOV, Unité information, Date Envoi AMF, Date Envoi Marché, Filename, Code Diffuseur, Société, ISO_CD_AMF, Site Internet, Siège Social, ISIN, Site Internet, Logo émetteur, identificationsociete_iso_cd_lei, Nom du fichier, Date de transmission, Titre du fichier, Type d'information, INF_COD_DIF, Type d'information, Liste de Diffusion, Langue, ISO_CD_MCH, ISO_CD_TKR, INF_UPG_REL, Statut de l'information, url_de_recuperation, timezone_uin_dat_amf, CAC40, name_cac40, Type d'information, Sous-type d'information, Type of information, Subtype of information, type d'informationn, type_d_information_ancien, type_d_information_nouveau, type_of_information_ancien, type_of_information_nouveau, code_isin_nom_sc, text, text, text, datetime, datetime, text, text, text, text, text, text, text, text, text, text, text, datetime, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text, text
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          identificationsociete_iso_cd_isi, identificationsociete_iso_nom_soc, nbre_enregistrements, code_isin_nom_sc, NA, NA, NA, NA, identificationsociete_iso_cd_isi, identificationsociete_iso_nom_soc, nbre_enregistrements, code_isin_nom_sc, text, text, int, text
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  codelov, type_d_information, sous_type_d_information, type_of_information, subtype_of_information, filename, type_informationn, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, NA, NA, NA, NA, NA, NA, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, NA, NA, NA, NA, NA, NA, >, codelov, Type d'information, Sous-type d'information, Type of information, Subtype of information, Filename, type d'informationn, text, text, text, text, text, text, text
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        identificationsociete_iso_cd_lei, identificationsociete_iso_nom_soc, nbre_enregistrements, NA, NA, NA, identificationsociete_iso_cd_lei, identificationsociete_iso_nom_soc, nbre_enregistrements, text, text, int
#>       metas.default.title             metas.default.description
#> 1 Sous-type d'information                                  <NA>
#> 2          Sociétés CAC40                                  <NA>
#> 3     Flux AMF (NEW PROD) <p>Bascule vers nouveau FTP <br/></p>
#> 4          Liste code ISI                                  <NA>
#> 5     Types d'information                                  <NA>
#> 6               Codes LEI                                  <NA>
#>   metas.default.theme metas.default.keyword metas.default.license
#> 1                  NA                    NA                    NA
#> 2                  NA                    NA                    NA
#> 3                  NA                    NA                    NA
#> 4                  NA                    NA                    NA
#> 5                  NA                    NA                    NA
#> 6                  NA                    NA                    NA
#>   metas.default.license_url metas.default.language
#> 1                        NA                     fr
#> 2                        NA                     fr
#> 3                        NA                     fr
#> 4                        NA                     fr
#> 5                        NA                     fr
#> 6                        NA                     fr
#>   metas.default.metadata_languages metas.default.timezone
#> 1                               fr                     NA
#> 2                             NULL                     NA
#> 3                               fr                     NA
#> 4                               fr                     NA
#> 5                               fr                     NA
#> 6                               fr                     NA
#>      metas.default.modified metas.default.modified_updates_on_metadata_change
#> 1 2018-11-13T08:50:28+00:00                                             FALSE
#> 2 2018-10-14T11:31:19+00:00                                             FALSE
#> 3 2018-08-09T16:27:24+00:00                                             FALSE
#> 4 2018-12-07T16:44:13+00:00                                             FALSE
#> 5 2018-11-13T08:50:28+00:00                                             FALSE
#> 6 2019-04-08T11:50:13+00:00                                             FALSE
#>   metas.default.modified_updates_on_data_change
#> 1                                         FALSE
#> 2                                         FALSE
#> 3                                         FALSE
#> 4                                         FALSE
#> 5                                         FALSE
#> 6                                         FALSE
#>       metas.default.data_processed metas.default.metadata_processed
#> 1 2021-04-15T06:23:47.189000+00:00 2021-04-15T06:23:48.218000+00:00
#> 2 2019-11-19T23:34:22.916000+00:00 2019-11-19T23:34:23.995000+00:00
#> 3 2025-09-21T14:06:05.916000+00:00 2025-09-21T14:06:21.966000+00:00
#> 4 2025-09-21T00:00:08.571000+00:00 2025-09-21T00:00:08.810000+00:00
#> 5 2023-02-22T08:54:32.324000+00:00 2023-02-22T08:54:32.361000+00:00
#> 6 2025-09-21T04:00:03.581000+00:00 2025-09-21T04:00:03.708000+00:00
#>   metas.default.geographic_reference metas.default.geographic_reference_auto
#> 1                           world_fr                                    TRUE
#> 2                           world_fr                                    TRUE
#> 3                           world_fr                                    TRUE
#> 4                           world_fr                                    TRUE
#> 5                           world_fr                                    TRUE
#> 6                           world_fr                                    TRUE
#>   metas.default.territory metas.default.geometry_types metas.default.bbox
#> 1                  France                           NA                 NA
#> 2                  France                           NA                 NA
#> 3                  France                           NA                 NA
#> 4                  France                           NA                 NA
#> 5                  France                           NA                 NA
#> 6                  France                           NA                 NA
#>   metas.default.publisher metas.default.references metas.default.records_count
#> 1                      NA                       NA                          20
#> 2                      NA                       NA                        1055
#> 3                      NA                       NA                      512382
#> 4                      NA                       NA                        4436
#> 5                      NA                       NA                          55
#> 6                      NA                       NA                         929
#>   metas.default.attributions metas.default.source_domain
#> 1                         NA                          NA
#> 2                         NA                          NA
#> 3                         NA                          NA
#> 4                         NA                          NA
#> 5                         NA                          NA
#> 6                         NA                          NA
#>   metas.default.source_domain_title metas.default.source_domain_address
#> 1                                NA                                  NA
#> 2                                NA                                  NA
#> 3                                NA                                  NA
#> 4                                NA                                  NA
#> 5                                NA                                  NA
#> 6                                NA                                  NA
#>   metas.default.source_dataset metas.default.shared_catalog
#> 1                           NA                           NA
#> 2                           NA                           NA
#> 3                           NA                           NA
#> 4                           NA                           NA
#> 5                           NA                           NA
#> 6                           NA                           NA
#>   metas.default.federated metas.default.parent_domain
#> 1                   FALSE                    datadila
#> 2                   FALSE                        <NA>
#> 3                   FALSE                    datadila
#> 4                   FALSE                    datadila
#> 5                   FALSE                    datadila
#> 6                   FALSE                    datadila
#>   metas.default.update_frequency
#> 1                             NA
#> 2                             NA
#> 3                             NA
#> 4                             NA
#> 5                             NA
#> 6                             NA
#>