get series infos from banque de france dataset catalog

retourne un data.frame d'information sur les series d'un dataset du catalogue banque de france

bdf_list_datasets(dataset = "OFC")

Arguments

dataset

le code du catalogue

Valeur de retour

data.frame

Examples

bdf_list_datasets(dataset = "OFC")
#>                                                 series_key
#> 1     OFC.Q.N.FR.W1.S12M.S1.N.A.LE.F._Z._Z.XDC._T.S.V.N._T
#> 2    OFC.Q.N.FR.W2.S12M.S13.N.A.LE.F._Z._Z.XDC._T.S.V.N._T
#> 3   OFC.Q.N.FR.W2.S12M.S12K.N.L.LE.F._Z._Z.XDC._T.S.V.N._T
#> 4    OFC.Q.N.FR.W2.S12M.S1V.N.L.LE.F._Z._Z.XDC._T.S.V.N._T
#> 5    OFC.Q.N.FR.W2.S12M.S13.N.L.LE.F._Z._Z.XDC._T.S.V.N._T
#> 6   OFC.Q.N.FR.W2.S12M.S12K.N.A.LE.F._Z._Z.XDC._T.S.V.N._T
#> 7  OFC.Q.N.FR.W2.S12M.S12M.N.NE.LE.F._Z._Z.XDC._T.S.V.N._T
#> 8    OFC.Q.N.FR.W2.S12M.S1V.N.A.LE.F._Z._Z.XDC._T.S.V.N._T
#> 9     OFC.Q.N.FR.W1.S12M.S1.N.L.LE.F._Z._Z.XDC._T.S.V.N._T
#> 10    OFC.Q.N.FR.W2.S12M.S1.N.N.LE.F._Z._Z.XDC._T.S.V.N._T
#> 11   OFC.Q.N.FR.W0.S12M.S1.N.L.LE.F5._Z._Z.XDC._T.S.V.N._T
#> 12   OFC.Q.N.FR.W2.S12M.S13.N.N.LE.F._Z._Z.XDC._T.S.V.N._T
#> 13    OFC.Q.N.FR.W1.S12M.S1.N.N.LE.F._Z._Z.XDC._T.S.V.N._T
#>                                                                                                                                                                   title_fr
#> 1                                                                                                                                        Actif vis-à-vis du reste du monde
#> 2                                                                                                                      Actif vis-à-vis des administrations publiques (APU)
#> 3                                                                                                           Passif vis-à-vis des institutions financières monétaires (IFM)
#> 4  Autres passifs  vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages (hors autres passifs nets)
#> 5                                                                                                                     Passif vis-à-vis des administrations publiques (APU)
#> 6                                                                                                            actif vis-à-vis des institutions financières monétaires (IFM)
#> 7                                                                                                                                                      Autres passifs nets
#> 8                                       Actif vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages
#> 9                                                                                                                                       Passif vis-à-vis du reste du monde
#> 10                                                                                                                             Actif net vis-à-vis de l'économie nationale
#> 11                                                                                                                           Actions et autres titres de propriété, actifs
#> 12                                                                                                                 Actif net vis-à-vis des administrations publiques (APU)
#> 13                                                                                                                                   Actif net vis-à-vis du reste du monde
#>                                                                                                                                                                                                          title_long_fr
#> 1                                                                                                                                        Autres sociétés financières, actif vis-à-vis du reste du monde, encours, euro
#> 2                                                                                                                      Autres sociétés financières, actif vis-à-vis des administrations publiques (APU), encours, euro
#> 3                                                                                                           Autres sociétés financières, passif vis-à-vis des institutions financières monétaires (IFM), encours, euro
#> 4  Autres sociétés financières, autres passifs  vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages (hors autres passifs nets), encours, euro
#> 5                                                                                                                     Autres sociétés financières, passif vis-à-vis des administrations publiques (APU), encours, euro
#> 6                                                                                                            Autres sociétés financières, actif vis-à-vis des institutions financières monétaires (IFM), encours, euro
#> 7                                                                                                                                                      Autres sociétés financières, autres passifs nets, encours, euro
#> 8                                      Autres sociétés financières, actif vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages , encours, euro
#> 9                                                                                                                                       Autres sociétés financières, passif vis-à-vis du reste du monde, encours, euro
#> 10                                                                                                                             Autres sociétés financières, actif net vis-à-vis de l'économie nationale, encours, euro
#> 11                                                                                                                                   Autres sociétés financières, actions et autres titres de propriété, encours, euro
#> 12                                                                                                                 Autres sociétés financières, actif net vis-à-vis des administrations publiques (APU), encours, euro
#> 13                                                                                                                                   Autres sociétés financières, actif net vis-à-vis du reste du monde, encours, euro
#>                                  title_en
#> 1                  Claims on nonresidents
#> 2            Claims on general government
#> 3  Liabilities on depository corporations
#> 4              Other domestic liabilities
#> 5       Liabilities on general government
#> 6       Claims on depository corporations
#> 7                       Other items (net)
#> 8                 Claims on other sectors
#> 9             Liabilities to nonresidents
#> 10                        Domestic claims
#> 11        Shares and other equity, assets
#> 12       Net claims on general government
#> 13                     Net Foreign assets
#>                                                                                                          title_long_en
#> 1                                                   Other Financial corporations, Claims on nonresidents, stocks, euro
#> 2                                             Other Financial corporations, Claims on general government, stocks, euro
#> 3                                   Other Financial corporations, Liabilities on depository corporations, stocks, euro
#> 4  Other Financial corporations, Other domestic liabilities (except those included in other items (net)), stocks, euro
#> 5                                        Other Financial corporations, Liabilities on general government, stocks, euro
#> 6                                        Other Financial corporations, Claims on depository corporations, stocks, euro
#> 7                                                        Other Financial corporations, Other items (net), stocks, euro
#> 8                                                  Other Financial corporations, Claims on other sectors, stocks, euro
#> 9                                              Other Financial corporations, Liabilities to nonresidents, stocks, euro
#> 10                                                         Other Financial corporations, Domestic claims, stocks, euro
#> 11                                         Other Financial corporations, Shares and other equity, assets, stocks, euro
#> 12                                        Other Financial corporations, Net claims on general government, stocks, euro
#> 13                                                      Other Financial corporations, Net Foreign assets, stocks, euro
#>    first_time_period_date last_time_period_date source_agency
#> 1              1995-12-31            2025-03-31           FR2
#> 2              1995-12-31            2025-03-31           FR2
#> 3              1995-12-31            2025-03-31           FR2
#> 4              1995-12-31            2025-03-31           FR2
#> 5              1995-12-31            2025-03-31           FR2
#> 6              1995-12-31            2025-03-31           FR2
#> 7              1995-12-31            2025-03-31           FR2
#> 8              1995-12-31            2025-03-31           FR2
#> 9              1995-12-31            2025-03-31           FR2
#> 10             1995-12-31            2025-03-31           FR2
#> 11             1995-12-31            2025-03-31           FR2
#> 12             1995-12-31            2025-03-31           FR2
#> 13             1995-12-31            2025-03-31           FR2
#>                                                                                                                                                                                                                                                                                                                                                                                   series_dimensions_and_values
#> 1     { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W1", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "A", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 2    { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S13", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "A", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 3   { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S12K", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "L", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 4    { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1V", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "L", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 5    { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S13", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "L", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 6   { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S12K", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "A", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 7  { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S12M", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "NE", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 8    { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1V", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "A", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 9     { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W1", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "L", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 10    { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "N", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 11   { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W0", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "L", "STO" : "LE", "INSTR_ASSET" : "F5", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 12   { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W2", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S13", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "N", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#> 13    { "FREQ" : "Q", "ADJUSTMENT" : "N", "REF_AREA" : "FR", "COUNTERPART_AREA" : "W1", "REF_SECTOR" : "S12M", "COUNTERPART_SECTOR" : "S1", "CONSOLIDATION" : "N", "ACCOUNTING_ENTRY" : "N", "STO" : "LE", "INSTR_ASSET" : "F", "MATURITY" : "_Z", "EXPENDITURE" : "_Z", "UNIT_MEASURE" : "XDC", "CURRENCY_DENOM" : "_T", "VALUATION" : "S", "PRICES" : "V", "TRANSFORMATION" : "N", "CUST_BREAKDOWN" : "_T" }
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     series_attributes_and_values
#> 1                                                                                                                                                                                                                                                                                                                                   {"TITLE": "Actif vis-à-vis du reste du monde", "DECIMALS": "0", "EXT_TITLE": "Claims on nonresidents", "NAT_TITLE": "Other Financial corporations, Claims on nonresidents, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif vis-à-vis du reste du monde, encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 2                                                                                                                                                                                                                                                                                   {"TITLE": "Actif vis-à-vis des administrations publiques (APU)", "DECIMALS": "0", "EXT_TITLE": "Claims on general government", "NAT_TITLE": "Other Financial corporations, Claims on general government, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif vis-à-vis des administrations publiques (APU), encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 3                                                                                                                                                                                                                                         {"TITLE": "Passif vis-à-vis des institutions financières monétaires (IFM)", "DECIMALS": "0", "EXT_TITLE": "Liabilities on depository corporations", "NAT_TITLE": "Other Financial corporations, Liabilities on depository corporations, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, passif vis-à-vis des institutions financières monétaires (IFM), encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 4  {"TITLE": "Autres passifs  vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages (hors autres passifs nets)", "DECIMALS": "0", "EXT_TITLE": "Other domestic liabilities", "NAT_TITLE": "Other Financial corporations, Other domestic liabilities (except those included in other items (net)), stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, autres passifs  vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages (hors autres passifs nets), encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 5                                                                                                                                                                                                                                                                       {"TITLE": "Passif vis-à-vis des administrations publiques (APU)", "DECIMALS": "0", "EXT_TITLE": "Liabilities on general government", "NAT_TITLE": "Other Financial corporations, Liabilities on general government, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, passif vis-à-vis des administrations publiques (APU), encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 6                                                                                                                                                                                                                                                     {"TITLE": "actif vis-à-vis des institutions financières monétaires (IFM)", "DECIMALS": "0", "EXT_TITLE": "Claims on depository corporations", "NAT_TITLE": "Other Financial corporations, Claims on depository corporations, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif vis-à-vis des institutions financières monétaires (IFM), encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 7                                                                                                                                                                                                                                                                                                                                                                         {"TITLE": "Autres passifs nets", "DECIMALS": "0", "EXT_TITLE": "Other items (net)", "NAT_TITLE": "Other Financial corporations, Other items (net), stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, autres passifs nets, encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 8                                                                                                                              {"TITLE": "Actif vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages", "DECIMALS": "0", "EXT_TITLE": "Claims on other sectors", "NAT_TITLE": "Other Financial corporations, Claims on other sectors, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif vis-à-vis des sociétés non financières et des ménages et institutions financières au but non lucratif au service des ménages , encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 9                                                                                                                                                                                                                                                                                                                       {"TITLE": "Passif vis-à-vis du reste du monde", "DECIMALS": "0", "EXT_TITLE": "Liabilities to nonresidents", "NAT_TITLE": "Other Financial corporations, Liabilities to nonresidents, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, passif vis-à-vis du reste du monde, encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 10                                                                                                                                                                                                                                                                                                                            {"TITLE": "Actif net vis-à-vis de l'économie nationale", "DECIMALS": "0", "EXT_TITLE": "Domestic claims", "NAT_TITLE": "Other Financial corporations, Domestic claims, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif net vis-à-vis de l'économie nationale, encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 11                                                                                                                                                                                                                                                                                                {"TITLE": "Actions et autres titres de propriété, actifs", "DECIMALS": "0", "EXT_TITLE": "Shares and other equity, assets", "NAT_TITLE": "Other Financial corporations, Shares and other equity, assets, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actions et autres titres de propriété, encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 12                                                                                                                                                                                                                                                                  {"TITLE": "Actif net vis-à-vis des administrations publiques (APU)", "DECIMALS": "0", "EXT_TITLE": "Net claims on general government", "NAT_TITLE": "Other Financial corporations, Net claims on general government, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif net vis-à-vis des administrations publiques (APU), encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}
#> 13                                                                                                                                                                                                                                                                                                                                  {"TITLE": "Actif net vis-à-vis du reste du monde", "DECIMALS": "0", "EXT_TITLE": "Net Foreign assets", "NAT_TITLE": "Other Financial corporations, Net Foreign assets, stocks, euro", "UNIT_MULT": "6", "TITLE_COMPL": "Autres sociétés financières, actif net vis-à-vis du reste du monde, encours, euro", "SOURCE_AGENCY": "FR2", "TIME_PER_COLLECT": "E"}